requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
python-dotenv==1.2.1
gunicorn==23.0.0
